fastapi-backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.21.1
python-dotenv==1.1.1
sqlmodel==0.0.8
psycopg2-binary==2.9.9
httpx==0.24.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
